represent value in satoshi instead of btc

+++ app/src/main/java/cz/skala/trezorwallet/data/entity/Transaction.kt
@@ -20,8 +20,8 @@ class Transaction(
         val blocktime: Long?,
         val confirmations: Int,
         val type: Type,
-        val value: Double,
-        val fee: Double,
+        val value: Long,
+        val fee: Long,
         val locktime: Int
 ) {
     enum class Type {
